make all_nan_classification_ids a property on AnalysisMixin
