+++ onecodex/viz/_heatmap.py
@@ -9,14 +9,6 @@ from onecodex.viz._primitives import (
 )
 
 
-def get_all_nan_classification_ids(df):
-    all_nan_classification_ids = []
-    for class_id, is_all_nan in df.isnull().all(1).items():
-        if is_all_nan:
-            all_nan_classification_ids.append(class_id)
-    return all_nan_classification_ids
-
-
 class VizHeatmapMixin(object):
     def plot_heatmap(
         self,
@@ -131,8 +123,7 @@ class VizHeatmapMixin(object):
             table_format="long",
             include_nans=True,
         )
-        # this won't work because we've already taken the top_n
-        all_nan_classification_ids = get_all_nan_classification_ids(df)
+        all_nan_classification_ids = self.all_nan_classification_ids
 
         if len(df["tax_id"].unique()) < 2:
             raise PlottingException(
